+++ apps/web/src/features/moim-create/ui/moim-form-fields.tsx
@@ -1,8 +1,7 @@
 "use client";
 
 import type { UseFormReturn } from "react-hook-form";
-import type { MoimCreateFormValues } from "@/features/moim-create/model/schema";
-import type { MoimFormState } from "@/features/moim-create/model/types";
+import type { MoimCreateFormValues, MoimFormState } from "@/entities/moim";
 import { FormActions } from "@/features/moim-create/ui/moim-form/form-actions";
 import { FormMoimSection } from "@/features/moim-create/ui/moim-form/form-moim-section";
 import { FormSpaceSection } from "@/features/moim-create/ui/moim-form/form-space-section";
