+++ apps/web/src/features/moim-create/ui/moim-form/form-space-section.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { Controller, type UseFormReturn } from "react-hook-form";
-import type { MoimCreateFormValues } from "@/features/moim-create/model/schema";
+import type { MoimCreateFormValues } from "@/entities/moim";
 import { FieldError } from "@/features/moim-create/ui/moim-form/form-field-error";
 import { FormLabel } from "@/features/moim-create/ui/moim-form/form-label";
 import { ThemeColorSelect } from "@/features/moim-create/ui/moim-form/theme-color-select";
