+++ apps/web/src/features/moim-create/ui/moim-form/theme-color-select.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { SelectBox } from "@ui/components/ui/selectbox";
-import { THEME_COLORS } from "@/features/moim-create/constants";
+import { THEME_COLORS } from "@/entities/moim";
 
 interface ThemeColorSelectProps {
   value?: string;
