+++ apps/web/src/features/moim-edit/ui/moim-edit-form.tsx
@@ -1,8 +1,8 @@
 "use client";
 
 import { useRouter } from "next/navigation";
+import type { MoimCreateFormValues } from "@/entities/moim";
 import { useMoimFormImageUpload } from "@/features/moim-create/hooks/use-moim-form-image-upload";
-import type { MoimCreateFormValues } from "@/features/moim-create/model/schema";
 import { MoimFormFields } from "@/features/moim-create/ui/moim-form-fields";
 import { useMoimEditForm } from "@/features/moim-edit/model/use-moim-edit-form";
